BitStream refactoring
