+++ Common/include/BitStream.hpp
