+++ Common/src/png/Type01Decoder.cpp
@@ -1,6 +1,6 @@
 #include "png/Type01Decoder.hpp"
-#include "png/BitStream.hpp"
 #include "png/Adler32.hpp"
+#include "BitStream.hpp"
 #include "Exception.hpp"
 
 #ifdef ENABLE_LOGS
